refact(managevenv): move `requirements` parsing into separate file
This way we can re-use the logic in multiple scripts.
